notebooks/requirements.txt
pandas>=2.0.0
clickhouse-driver>=0.2.0
sqlalchemy>=2.0.0
papermill>=2.4.0
ipykernel>=6.0.0
clickhouse-sqlalchemy>=0.2.4
pyyaml>=6.0.0
python-dotenv>=1.0.0
